requirements.txt
absl-py==1.4.0
anyio==3.7.0
asgiref==3.7.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Django==4.2.2
exceptiongroup==1.1.1
fastapi==0.97.0
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-api-core==2.11.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
idna==3.4
image==1.5.33
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax==0.4.12
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
typing-extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0
